Change way containers are created
The user can now choose the name of the container, as the name is
no indication of what kind of container is made. Every container
is now dual-stacked with IP and XIA.

This patch also lets users choose the range of containers to be
created, so containers can be added at a later time.